requirements.txt
requests==2.31.0
pymongo==4.6.1
minio==7.2.5
Pillow==10.2.0
python-dateutil==2.8.2
dnspython==2.6.1
pandas==2.2.1